// src/generated/prisma/enums.ts
export const UserRole = {
  CUSTOMER: 'CUSTOMER',
  SELLER: 'SELLER',
  CS: 'CS',
  MODERATOR: 'MODERATOR',
  FINANCE: 'FINANCE',
  ADMIN: 'ADMIN',
  SUPER: 'SUPER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  SUSPENDED: 'SUSPENDED',
  SELLING_RESTRICTED: 'SELLING_RESTRICTED',
  WITHDRAWAL_HOLD: 'WITHDRAWAL_HOLD',
  BANNED: 'BANNED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]